Automatically disabling the plugin should be enough. No need for other changes
