+++ spec/requests/sitemap_controller_spec.rb
@@ -2,10 +2,10 @@
 
 require 'rails_helper'
 
-RSpec.describe DiscourseSitemap::SitemapController do
+RSpec.describe 'DiscourseSitemap::SitemapController', if: defined?(DiscourseSitemap::SitemapController) do
   describe '#default' do
     before do
-      DiscourseSitemap::Sitemap.create!(name: '1', last_posted_at: Time.now)
+      Sitemap.create!(name: '1', last_posted_at: Time.now)
     end
 
     it 'does not fail then page is a string starting with a number' do
